use go-generate to generate LangCodeInfo
