+++ language_codes.csv
@@ -1,0 +1,184 @@
+aa,"Afar","Afaraf"
+ab,"Abkhazian","аҧсуа бызшәа; аҧсшәа"
+ae,"Avestan","avesta"
+af,"Afrikaans","Afrikaans"
+ak,"Akan","Akan"
+am,"Amharic","አማርኛ"
+an,"Aragonese","aragonés"
+ar,"Arabic","العربية"
+as,"Assamese","অসমীয়া"
+av,"Avaric","авар мацӀ; магӀарул мацӀ"
+ay,"Aymara","aymar aru"
+az,"Azerbaijani","azərbaycan dili"
+ba,"Bashkir","башҡорт теле"
+be,"Belarusian","беларуская мова"
+bg,"Bulgarian","български език"
+bh,"Bihari languages","भोजपुरी"
+bi,"Bislama","Bislama"
+bm,"Bambara","bamanankan"
+bn,"Bengali","বাংলা"
+bo,"Tibetan","བོད་ཡིག"
+br,"Breton","brezhoneg"
+bs,"Bosnian","bosanski jezik"
+ca,"Catalan; Valencian","català; valencià"
+ce,"Chechen","нохчийн мотт"
+ch,"Chamorro","Chamoru"
+co,"Corsican","corsu; lingua corsa"
+cr,"Cree","ᓀᐦᐃᔭᐍᐏᐣ"
+cs,"Czech","čeština; český jazyk"
+cu,"Church Slavic; Old Slavonic; Church Slavonic; Old Bulgarian; Old Church Slavonic","ѩзыкъ словѣньскъ"
+cv,"Chuvash","чӑваш чӗлхи"
+cy,"Welsh","Cymraeg"
+da,"Danish","dansk"
+de,"German","Deutsch"
+dv,"Divehi; Dhivehi; Maldivian","ދިވެހި"
+dz,"Dzongkha","རྫོང་ཁ"
+ee,"Ewe","Eʋegbe"
+el,"Greek, Modern (1453-)","ελληνικά"
+en,"English","English"
+eo,"Esperanto","Esperanto"
+es,"Spanish; Castilian","español; castellano"
+et,"Estonian","eesti; eesti keel"
+eu,"Basque","euskara; euskera"
+fa,"Persian","فارسی"
+ff,"Fulah","Fulfulde; Pulaar; Pular"
+fi,"Finnish","suomi; suomen kieli"
+fj,"Fijian","vosa Vakaviti"
+fo,"Faroese","føroyskt"
+fr,"French","français; langue française"
+fy,"Western Frisian","Frysk"
+ga,"Irish","Gaeilge"
+gd,"Gaelic; Scottish Gaelic","Gàidhlig"
+gl,"Galician","galego"
+gn,"Guarani","Avañeẽ"
+gu,"Gujarati","ગુજરાતી"
+gv,"Manx","Gaelg; Gailck"
+ha,"Hausa","(Hausa) هَوُسَ"
+he,"Hebrew","עברית"
+hi,"Hindi","हिन्दी; हिंदी"
+ho,"Hiri Motu","Hiri Motu"
+hr,"Croatian","hrvatski jezik"
+ht,"Haitian; Haitian Creole","Kreyòl ayisyen"
+hu,"Hungarian","magyar"
+hy,"Armenian","Հայերեն"
+hz,"Herero","Otjiherero"
+ia,"Interlingua (International Auxiliary language Association)","Interlingua"
+id,"Indonesian","Bahasa Indonesia"
+ie,"Interlingue; Occidental","Interlingue; Occidental"
+ig,"Igbo","Asụsụ Igbo"
+ii,"Sichuan Yi; Nuosu","ꆈꌠ꒿ Nuosuhxop"
+ik,"Inupiaq","Iñupiaq; Iñupiatun"
+io,"Ido","Ido"
+is,"Icelandic","Íslenska"
+it,"Italian","italiano"
+iu,"Inuktitut","ᐃᓄᒃᑎᑐᑦ"
+ja,"Japanese","日本語(にほんご)"
+jv,"Javanese","basa Jawa"
+ka,"Georgian","ქართული"
+kg,"Kongo","Kikongo"
+ki,"Kikuyu; Gikuyu","Gĩkũyũ"
+kj,"Kuanyama; Kwanyama","Kuanyama"
+kk,"Kazakh","қазақ тілі"
+kl,"Kalaallisut; Greenlandic","kalaallisut; kalaallit oqaasii"
+km,"Central Khmer","ខ្មែរ; ខេមរភាសា; ភាសាខ្មែរ"
+kn,"Kannada","ಕನ್ನಡ"
+ko,"Korean","한국어; 조선어"
+kr,"Kanuri","Kanuri"
+ks,"Kashmiri","कश्मीरी; كشميري‎"
+ku,"Kurdish","Kurdî; كوردی‎"
+kv,"Komi","коми кыв"
+kw,"Cornish","Kernewek"
+ky,"Kirghiz; Kyrgyz","Кыргызча; Кыргыз тили"
+la,"Latin","latine; lingua latina"
+lb,"Luxembourgish; Letzeburgesch","Lëtzebuergesch"
+lg,"Ganda","Luganda"
+li,"Limburgan; Limburger; Limburgish","Limburgs"
+ln,"Lingala","Lingála"
+lo,"Lao","ພາສາລາວ"
+lt,"Lithuanian","lietuvių kalba"
+lu,"Luba-Katanga","Tshiluba"
+lv,"Latvian","latviešu valoda"
+mg,"Malagasy","fiteny malagasy"
+mh,"Marshallese","Kajin M̧ajeļ"
+mi,"Maori","te reo Māori"
+mk,"Macedonian","македонски јазик"
+ml,"Malayalam","മലയാളം"
+mn,"Mongolian","монгол"
+mr,"Marathi","मराठी"
+ms,"Malay","bahasa Melayu; بهاس ملايو‎"
+mt,"Maltese","Malti"
+my,"Burmese","ဗမာစာ"
+na,"Nauru","Ekakairũ Naoero"
+nb,"Bokmål, Norwegian; Norwegian Bokmål","Norsk bokmål"
+nd,"Ndebele, North; North Ndebele","isiNdebele"
+ne,"Nepali","नेपाली"
+ng,"Ndonga","Owambo"
+nl,"Dutch; Flemish","Nederlands; Vlaams"
+nn,"Norwegian Nynorsk; Nynorsk, Norwegian","Norsk nynorsk"
+no,"Norwegian","Norsk"
+nr,"Ndebele, South; South Ndebele","isiNdebele"
+nv,"Navajo; Navaho","Diné bizaad; Dinékʼehǰí"
+ny,"Chichewa; Chewa; Nyanja","chiCheŵa; chinyanja"
+oc,"Occitan (post 1500); Provençal","occitan; lenga dòc"
+oj,"Ojibwa","ᐊᓂᔑᓈᐯᒧᐎᓐ"
+om,"Oromo","Afaan Oromoo"
+or,"Oriya","ଓଡ଼ିଆ"
+os,"Ossetian; Ossetic","ирон æвзаг"
+pa,"Panjabi; Punjabi","ਪੰਜਾਬੀ; پنجابی‎"
+pi,"Pali","पाऴि"
+pl,"Polish","język polski; polszczyzna"
+ps,"Pushto; Pashto","پښتو"
+pt,"Portuguese","português"
+qu,"Quechua","Runa Simi; Kichwa"
+rm,"Romansh","rumantsch grischun"
+rn,"Rundi","Ikirundi"
+ro,"Romanian; Moldavian; Moldovan","limba română"
+ru,"Russian","русский язык"
+rw,"Kinyarwanda","Ikinyarwanda"
+sa,"Sanskrit","संस्कृतम्"
+sc,"Sardinian","sardu"
+sd,"Sindhi","सिन्धी; سنڌي، سندھی‎"
+se,"Northern Sami","Davvisámegiella"
+sg,"Sango","yângâ tî sängö"
+si,"Sinhala; Sinhalese","සිංහල"
+sk,"Slovak","slovenčina; slovenský jazyk"
+sl,"Slovenian","slovenski jezik; slovenščina"
+sm,"Samoan","gagana faa Samoa"
+sn,"Shona","chiShona"
+so,"Somali","Soomaaliga; af Soomaali"
+sq,"Albanian","Shqip"
+sr,"Serbian","српски језик"
+ss,"Swati","SiSwati"
+st,"Sotho, Southern","Sesotho"
+su,"Sundanese","Basa Sunda"
+sv,"Swedish","Svenska"
+sw,"Swahili","Kiswahili"
+ta,"Tamil","தமிழ்"
+te,"Telugu","తెలుగు"
+tg,"Tajik","тоҷикӣ;toğikī;تاجیکی‎"
+th,"Thai","ไทย"
+ti,"Tigrinya","ትግርኛ"
+tk,"Turkmen","Türkmen; Түркмен"
+tl,"Tagalog","Wikang Tagalog; ᜏᜒᜃᜅ᜔ ᜆᜄᜎᜓᜄ᜔"
+tn,"Tswana","Setswana"
+to,"Tonga (Tonga Islands)","faka Tonga"
+tr,"Turkish","Türkçe"
+ts,"Tsonga","Xitsonga"
+tt,"Tatar","татар теле; tatar tele"
+tw,"Twi","Twi"
+ty,"Tahitian","Reo Tahiti"
+ug,"Uighur; Uyghur","Uyƣurqə; ئۇيغۇرچە‎"
+uk,"Ukrainian","українська мова"
+ur,"Urdu","اردو"
+uz,"Uzbek","O‘zbek; Ўзбек; أۇزبېك‎"
+ve,"Venda","Tshivenḓa"
+vi,"Vietnamese","Tiếng Việt"
+vo,"Volapük","Volapük"
+wa,"Walloon","walon"
+wo,"Wolof","Wollof"
+xh,"Xhosa","isiXhosa"
+yi,"Yiddish","ייִדיש"
+yo,"Yoruba","Yorùbá"
+za,"Zhuang; Chuang","Saɯ cueŋƅ; Saw cuengh"
+zh,"Chinese","中文(Zhōngwén); 汉语; 漢語"
+zu,"Zulu","isiZulu"
